Refactor - moved KorAllyToken to shared area
