+++ Mage.Sets/src/mage/sets/oathofthegatewatch/OathOfGideon.java
@@ -44,6 +44,7 @@ import mage.game.Game;
 import mage.game.events.EntersTheBattlefieldEvent;
 import mage.game.events.GameEvent;
 import mage.game.permanent.Permanent;
+import mage.game.permanent.token.KorAllyToken;
 
 /**
  *
